[FIX] {mail, mass_mailing}/saas~15.2: ensure valid trace / activities
Activities and mailing traces are linked to documents records through a model
and a many2one reference field. Those fields are required but are implemented
using integer fields. This means writing 0 is actually a valid value even if
technically 'required'. We added constraints to ensure res_id is not void.

In this commit we update migration so that invalid records are removed.
As they cannot be managed and link to access errors we should not try to
keep them or attach to some random record.

Mailing traces have been updated at odoo/upgrade@ffd0e770d88493841324563bcf8a8042eabfed0f . However as in between due to
0 being valid for a required field new wrong traces may have been generated
we clean the table again.

Source: internal feedback about records with 0 as res_id

See community odoo/odoo#81292
See enterprise odoo/enterprise#23652

Task-2694133

closes odoo/upgrade#3106

Signed-off-by: Christophe Simonis <chs@odoo.com>