[FIX] mrp_workorder/saas~13.4: duplicate specific quality point
In v13 an operation could be shared between multiple boms. Quality
points linked to this operation are filtered depending on the finished
product to get the relevant ones during a production.

In v14, each bom has its own operations as well as its own quality
points. The migration script that duplicate the quality points after
duplicating the operations did not made this filter on the finished
product.

This miss implies to have all the quality points (and thus all the
quality check at production creation) on every BoMs, even the ones with
another product to produce.

Request 23465
opw-2641908

closes odoo/upgrade#3120

Signed-off-by: Christophe Simonis <chs@odoo.com>