[FIX] website@saas~14.4: correct website's domain
Commit d3080ea872e9 was in charge of cleaning the website domain:
- Removing the trailing slashes
- Adding the scheme
It was basically just mimicking [1] which was doing that when writing `domain`
on a `website`, in order to ensure already existing DBs get that improvement
and `domain` could be reliable.

But it was not correctly done, it was missing the check to ensure the domain
was set, before prefixing it with a scheme, as we do in community.

Step to reproduce:
  - In 14.0, create a website with no domain
  - Migrate the DB to 15.0
  - Website domain is now `https://`, which is wrong

[1]: https://github.com/odoo/odoo/commit/042c95b0219bb0aa13e73385e092fa76ff1a1b0a

task-2720185

closes odoo/upgrade#3130

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>