[FIX] project/saas~14.5.1.2: Maintain assignees in project task
The condition to maintain the assigned 'user_id' for 'project_task'
by maintaining 'project_task_user_rel' where the stages created by
user had fold as null in version 14.0. This weren't considered to be
inserted to this table and due to that the users weren't assigned to
specific project were taken into consideration.

closes odoo/upgrade#3147

Signed-off-by: Christophe Simonis <chs@odoo.com>