[FIX] account/saas~14.3.1.1: Guess companies' country
Addresses Traceback Group:
https://upgrade.odoo.com/web#id=134&action=178&model=\
upgrade.request.traceback.group&view_type=form&cids=1&menu_id=107

Each company needs to have its fiscal country (`account_tax_fiscal_country_id`)
set, but that information is not always included in the `res_company` table
(the country might be deduced from its Chart of Accounts or l10n modules, see
https://github.com/odoo/upgrade/pull/1264,
https://github.com/odoo/upgrade/pull/2322).

We thus turn to the rest of the available information for an informed guess of
the company's country, by considering the following clue sources:

A) The CoA of the company
B) The company's "partner" record's country
C) Whether the company's name includes the country
(e.g. 'Tech Solutions Australia')
D) Whether the company has a country-specific timezone
(e.g. 'Europe/Brussels')
E) Whether the company's (mobile) phone number includes a country phone code
(e.g. '+32 01234567')
F) The company's currency
(e.g. 'Yen')
G) Whether  company's website/email addresses (in that order) are registered
under a country Top Level Domain (e.g. 'support@steakinc.ar')
H) The company's language locale

To achieve that, we look at both the company and its entry in the partner
table, in that order, until we find a clue.
If a clue yields more than one possible country, an attempt is made to see if
there's only a single "main" country which could be chosen ("main" countries
are independent states, so among UK, Guernsey and Jersey the former would be
preferred). The rationale is that dependent territories are often much smaller
and statistically less likely than independent states.

Finally, if no clue is found, that company's language is consulted -
specifically, the country code at the end of the language code.
Often this is just the default 'en_US', so if no other clue helped,
the company will be set to 'United States'.

This fix is not guaranteed to set all companies correctly; it merely looks at
the more immediate hints and makes a best effort to choose a viable country.
The user is still encouraged to check and correct any countries guessed after
the upgrade is complete.

Tests have also been added for the clues:
For each clue, a new company is created with the clue's relevant fields filled
accordingly. In some cases, some "fake" information is also added, to test
whether the clues operate in the manner and order they're specified. Those
tests are marked by stating "(CountryA - not CountryB)" in the comment in the
test file.