[FIX] crm/saas~14.4: fix translation on arch_db
Fix for #3076.
Remove the translation of failing terms in the arch_db of
`crm.crm_lead_view_form` to force them to be udpdated during the
upgrade.

Issue:
While upgrading this module, the translations that correctly translate
the domain on `crm.crm_lead_view_form` that use `calendar_event_count`
https://github.com/odoo/odoo/blob/8a7d41c99ec493d86166708552c8df51eaa268c2/addons/crm/views/crm_lead_views.xml#L34-L35
are not updated.

This happens because `_sync_terms_translation`[1] creates the term for
non English languages. Then when the PO file is loaded, since the term
is already there and `override` is `False` by default[2,3] it does not get
the updated translation. Even if it is correctly set on the language.
This prevents us to fix the issue by updating the translation on the PO
file.

[1] https://github.com/odoo/odoo/blob/704a47a520c789906b7e98ef48e5cb4f6592a13a/odoo/addons/base/models/ir_translation.py#L438-L441
[2] https://github.com/odoo/odoo/blob/20888055d4271bf3cf9e7bc3d42150a1f72e4495/odoo/modules/loading.py#L230-L231
[3] https://github.com/odoo/odoo/blob/704a47a520c789906b7e98ef48e5cb4f6592a13a/odoo/addons/base/models/ir_translation.py#L97-L136

The issue can be reproduced in standard:
* Create a clean DB in 14.0 with only CRM installed
* Set the language to `zh_CN`
* Create a dummy Lead
* On 15.0, update the PO file `crm/i18n/zh_CN.po` and add the correct
  translation using `calendar_event_count` on the domain
* Update from 14.0 to 15.0

Then if we open CRM > Lead form view we get a traceback on the
frontend.

closes odoo/upgrade#3165

Signed-off-by: Christophe Simonis <chs@odoo.com>