[FIX] website/saas~14.5: remove theme sale/blog bridge modules, again
Some DBs that got migrated in 15.0 are left with `_blog` and/or `_sale` theme
bridge module in the `ir_module_module` table, despite the module not existing
anymore. Indeed, those were deleted in versions 12.0 to 13.0, depending of
which one.

Since the migration script to remove the module was not merged at the same time
as the module removal in the `design-themes` repo, some DBs were not correctly
migrated and are left with those bridge modules entries.

Since 15.0, more than just being wrong, it raises an error in the website
configurator which then tries to read those module manifest files, which don't
exists anymore. (That code will be patched to avoid breaking)

Here is a summary of what happened:
1. theme_graphene_blog
   - Aug 28, 2020: Removed https://github.com/odoo/design-themes-old/pull/346
   - Sep 16, 2020: Upgrade script added 3 weeks later https://github.com/odoo/upgrade/pull/1684

2. theme_avantgarde_blog
   - Feb 5, 2020: Removed https://github.com/odoo/design-themes-old/pull/198
   - Dec 1, 2020: Upgrade script added 10 months later https://github.com/odoo/upgrade/pull/1981
   - opw-2728219

3. _sale extension of [Anelusia, Kea, Kiddo, Loftspace, Monglia, Notes]
   - Aug 28, 2020: Removed https://github.com/odoo/design-themes-old/pull/346
   - Sep 16, 2020: Upgrade script added 3 weeks later https://github.com/odoo/upgrade/pull/1684

4. _sale extension of [artists, beauty, bookstore, odoo_experts, orchid,
   real_estate, vehicle, yes]
   - Sep 20, 2019: Removed https://github.com/odoo/design-themes-old/pull/170
   - Dec  1, 2020: Upgrade script added over a year later https://github.com/odoo/upgrade/pull/1981
     Jul 14, 2020: [artists, odoo_experts, real_estate] were already removed by
                   script with https://github.com/odoo/upgrade/pull/1244
   - opw-2725887

Removing again those theme bridge modules will ensure DBs are now cleaned,
starting from 15.0.

Note that points 1 and 3 should not be problematic as the upgrade scripts were
committed before the release of version 14.

opw-2728219
opw-2725887

closes odoo/upgrade#3187

Signed-off-by: Christophe Simonis <chs@odoo.com>