[FIX] product/saas~14.5: compute type from detailed_type
How to reproduce:
* Delete `point_of_sale.product_product_tip`
* Add a `ir.defaults` for `product.template.type` that is not `consu`
* Upgrade

Result:
`_constrains_detailed_type` fails because when trying to recreate the
data, neither `type` or `detailed_type` is provided. Therefore, the
default values are set:
* `type` is the value set in the `ir.defaults`
* `detailed_type` is the value set on the field: `consu`

By renaming the field, the `ir.defaults` and the default set on the
field are both affecting the same field, hence solving the issue.

opw-2728404
upgrade-94438

closes odoo/upgrade#3190

Signed-off-by: Christophe Simonis <chs@odoo.com>