[IMP] fetchmail_gmail, google_gmail/saas~15.2: use new field for gmail
Purpose
=======
In stable, we used a field `use_google_gmail_service` to define a mail
server which use gmail authentication. But in master, we wanted to use
the field `server_type` for the incoming mail server and
`smtp_authentication` for the outgoing mail server.

Task-2170676
See odoo/odoo/pull/83413

closes odoo/upgrade#3199

Related: odoo/odoo#83413
Signed-off-by: Christophe Simonis <chs@odoo.com>