[IMP] survey/saas~15.3: remove column_nb from the survey_question model
The "column choice" is a remainder of a previous implementation in which users
could decide how the survey would look like.

As the field is not used anymore and some customers are wondering why it does
not work, we have decided to remove it.

PR: https://github.com/odoo/odoo/pull/83645
Task 2730409

closes odoo/upgrade#3231

Related: odoo/odoo#83645
Signed-off-by: Christophe Simonis <chs@odoo.com>