[FIX] base/saas~12.5: make PG performance more predictable
For unknown reasons PG has a high variability in performance when
removing the field image_small. It has been observed on DBs with
millions of partners that the removal of this field gets stuck for hours
on parallel queries like
`DELETE FROM ir_attachment WHERE model='res_partner' AND id BETWEEN ...`
while, under unknown circumstances, it takes just a few seconds.
We add here a commit call with the hope of making the processing more
predictable (comparison below).

Since this behaviour is non deterministic we cannot guarantee this will
work on all cases.

Original:
```
2022-02-03 15:14:10,163 4545 INFO 102725_afu_13.0 odoo.modules.migration: module base: Running migration [>saas~12.5.1.3] pre-21-images
2022-02-03 16:02:40,515 4545 INFO 102725_afu_13.0 odoo.upgrade.util.pg: [7.58%] 32/422 queries processed in 0:47:34.836146
```

With `commit` call:
```
2022-02-04 17:57:05,995 4553 INFO 102725_afu_13.0 odoo.modules.migration: module base: Running migration [>saas~12.5.1.3] pre-21-images
2022-02-04 17:58:21,333 4553 INFO 102725_afu_13.0 odoo.upgrade.util.pg: [100.00%] 422/422 queries processed in 0:00:15.408998
2022-02-04 17:58:27,238 4553 INFO 102725_afu_13.0 odoo.modules.migration: module base: Running migration [>saas~12.5.1.3] pre-30-data
```

closes odoo/upgrade#3229

Signed-off-by: Christophe Simonis <chs@odoo.com>