[IMP] {base,stock_barcode}/saas~15.3: remove picking form barcode scan
We are removing barcode scanning from stock.picking forms since its
outdated and we no longer want to maintain it. Therefore we remove the
views/inherits/wizards related to it. This also includes removing the
module `delivery_barcode` which is made obsolete from this change.

Task: 2039720
ENT PR: odoo/enterprise#24145

closes odoo/upgrade#3230

Signed-off-by: Christophe Simonis <chs@odoo.com>