[ADD] product/0.0.0: ensure demo product exists
Since [1], adding exclusions on templates deletes the autogenerated
variants.
Unfortunately, this is what happens with the demo data of the `sale`
module, which deletes demo data of the `product` module. Moreover, the
`product` module demo data are written in a way that expects this record
to exist when updating.
This is not a real problem, but it generates a lot of useless warnings
in the log of the CI platforms.
Avoid this by re-creating the missing record.

[1] https://github.com/odoo/odoo/commit/5ff4eb022757d7a522ccb975f8eb64053ef9780b

closes odoo/upgrade#3262

Signed-off-by: Christophe Simonis <chs@odoo.com>