[FIX] account/saas~14.5: Creation of zero percent tax line
see odoo/odoo@65b5ec4ad65f798b1d647978b5b42a49e760139a

Tax lines for 0% tax were not created prior to 15.0 because they were
considered noisy journal items. With 15.0 and the tax details being
used to compute the tax report, these missing lines are now an issue as
we are not able to properly display the 0% tax in the report.

In order to fix this incorrect behavior, we will create the tax lines
for 0% tax during the migration to v15.

Task id #2761930

closes odoo/upgrade#3276

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>