[FIX] util.remove_inherit_from_model: do not delete in parallel
Disabling constraints has the nasty effect of not executing their
`ON DELETE` action, which may forbid the re-addition of the constraint.

Moreover, the `ON DELETE` constraints can be indirect (through another
table).

However, now that we keep the constraints, we cannot explode the query
anymore.

original issue:
opw-2721509
upg-136692
Fixed by #3272

inability to re-add constraint issue:
upg-164009

Part-of: odoo/upgrade#3300