[IMP] account/tests/test_amounts.py
Due to interactions with other tests and upgrades, some conditions on
version and companies were made.
As newly posted entries can be added by demo data like it is done by the
new module `l10n_us_1099` added in version 15.0 [1], we need to ignore
the newly posted entries added after the start of the upgrade.

As real-world databases are not supposed to have demo data, this date
check is only done when run in CI.

With this date check in place, the explicit check against the existence
of the `TestTagNoInvert` test data can be removed.

Also, refactor the verification of the version the test runs on.

As we only run the test against databases >= saas~13.4, the filter on
the companies can be removed as it was added to handle a saas~12.4 test.

[1]: odoo/enterprise@3fb0c7f49c800a687a014393b9e73e1d4631eacc

closes odoo/upgrade#3307

Signed-off-by: Christophe Simonis <chs@odoo.com>