[IMP] sale/saas~11.2: only update when necessary
is_expense is created with a default "FALSE", then only update the 
"TRUE" ones seems more efficient.
Udating twice a record is slower the second time within a transaction, 
so a cr.commit betwen the 2 updates can improve performance, and the 
cr.commit is a side effect of the paallel_execute.

closes odoo/upgrade#3313

Signed-off-by: Christophe Simonis <chs@odoo.com>