[IMP] analytic/saas~15.2: remove translate from aal name
`account.analytic.line` `name` has been set as translatable in saas-15.1.
In hindsight, we can say it was clearly a bad idea as it introduced
complexity and confusion. This is why we are reverting it.

This commit updates the `name` of the `account.analytic.line` records
for which a translation has been done. It uses the content of the latest
created translation.

This commit also removes the translations from the database.

Related PR: odoo/odoo#86124

closes odoo/upgrade#3333

Signed-off-by: Christophe Simonis <chs@odoo.com>