[FIX] base/saas~12.3.1.0: fix possible deadlock
If decimal_precision gets a NOT NULL columnd it may deadlock. So far
there is custom module adding display_digits column that has caused at
least two deadlock. This commit is to prevent this to happen in the
future. The two cases so far have been solved as specific fixes.

Full details from KangOl:
```
At the end of the update, the method _check_removed_columns is called on each model.
This method is responsible of removing the NOT NULL constraints on unknown  fields.

The list of know field is get by filtering the stored fields with a column definition.
https://github.com/odoo/odoo/blob/336c9aab3619abb5bfbeb07f87ae014facdefb0b/odoo/models.py#L2344-L2345

On fields.Float, the column_type attribute is a property that will read the digits property that, in case of decimal precision, will create a new (lazy) cursor to read the value in db.

This is not a problem in most cases.
However, in this database, there is a custom module smile_decimal_precision (I know, it starts smelling) that add a new stored field on decimal.precision. And you know what? This field is required (NOT NULL).

So, when calling the _check_removed_columns method on the decimal.precision model, the NOT NULL constraint on this field is removed, which lock the decimal_precision table.

The code then call the _check_removed_columns method on the mrp.bom model, which will try to read the decimal precision of float fields in a new cursor.

And kaboom 💥 ! Deadlock.

The solution is to create a specific init- script to drop the NOT NULL constraint on this field.
ALTER TABLE decimal_precision ALTER COLUMN display_digits DROP NOT NULL;
```

upg-165730
upg-147140

closes odoo/upgrade#3342

Signed-off-by: Christophe Simonis <chs@odoo.com>