[IMP] project_timesheet_forecast/saas~15.3 update percentage_hours
The field `percentage_hours` in `planning.slot`, which was previously
unused, is now used for the planning slots progress bar.

However, it was computed on a scale of 0 to 1, whereas the progress bar
needs a number from 0 to 100.

This commit multiplies all the previous values of this field by 100.

See https://github.com/odoo/enterprise/pull/23342

Task-2664116

closes odoo/upgrade#3344

Related: odoo/enterprise#23342
Signed-off-by: Christophe Simonis <chs@odoo.com>