[FIX] *: `fix_fk_allowed_cascade` and `no_model_data_delete` without upgrading base
Because of the version change in odoo/odoo@2cbc87581b465a2d43ed4054b3fb1a8edcfe5d6e
saas-15.1 databases on the saas attempting to upgrade `account` alone
(on the saas, not on the upgrade platform)
ends up with the below exception:
```
File "/home/odoo/src/odoo/saas-15.1/odoo/addons/base/maintenance/migrations/account/0.0.0/pre-fix-fk.py", line 5, in migrate
util.ENVIRON["fix_fk_allowed_cascade"].append(("account_partial_reconcile", "debit_move_id"))
...
KeyError: 'fix_fk_allowed_cascade'
```

It's happening because the variable `fix_fk_allowed_cascade` is defined in the base module
in the upgrade script
https://github.com/odoo/upgrade/blob/bb78f9f6b903fed79ec149f5f887aa03217c24e0/migrations/base/0.0.0/pre-fix-fk.py#L9
which is not executed because `account` is being upgraded alone.
This is not occurring on the upgrade platform as all modules are upgraded,
therefore including base, therefore executing this script.
It wouldn't occur either if the upgrade scripts were not deployed on the saas,
but this is the case.
One could wonder if we shouldn't remove the upgrade scripts from the saas platform servers
as they are unused on it,
but anyway it's not a bad thing to support this case of upgrading one module alone,
without the base module included.

opw-2793725

closes odoo/upgrade#3361

Signed-off-by: Christophe Simonis <chs@odoo.com>