[FIX] base/0.0.0: rename non-conforming custom fields
Custom field names should start with a 'x_' and must be only
characters, digits, and underscores.
https://github.com/odoo/odoo/blob/9dcfae2234b1f45da1837c519b54926522cfa070/odoo/addons/base/models/ir_model.py#L566-L574

The field name constraint is not taking effect when web_studio is installed
since v15, and is causing issues when 'reflect_fields' is called on upgrade
due to the possible presence of spaces in the field names and the
sql constraint 'name_nospace' on ir_model_data table.
https://github.com/odoo/enterprise/blob/1d2da8a1498dc418fde5aa3e0491ab6d36920222/web_studio/models/ir_model.py#L591

Only parts of the constraint on custom field names can be automatically fixed
(prefix and spaces), other non standard characters if present are not replaced
by other characters, and for now it does not harm the name_nospaces constraint.

upg-168725

closes odoo/upgrade#3372

Signed-off-by: Christophe Simonis <chs@odoo.com>