[FIX] account/saas~12.4: fix matching bill price diff lines
When a Purchase Order does not match the Vendor Bill prices the lines
are not properly matched. This causes that on
https://github.com/odoo/upgrade/blob/32fe09b94572d910279bf5920a0aab9f8b9d00e3/migrations/purchase_stock/saas%7E12.4.1.2/end-10-account-pocalypse.py#L12-L35
the price difference is not properly reflected after migration.

If we try to migrate a test DB with a price difference we get good
results, as long as the PO have only one product. More precisely as long
as the associated invoice has only one debit line. They are matched with
the rule `{"same_account", "same_analytic_account", "only_one_line"}`.

For client DBs there are cases where the PO and BILL contains multiple
products for which the price was changed (increased or decreased) on
several of them. This won't allow the previous rule to match.

opw-2792468

closes odoo/upgrade#3371

Signed-off-by: Christophe Simonis <chs@odoo.com>