[FIX] base/saas~14.3: unmigrable web_editor assets
The web editor allow the users to customize some JS, CSS and XML assets.
Some of them are not meant to get customized through the web editor and
in this situation it warns the users that this is not the proper way to
customize those.

Before this commit
When a user had ignored this warning and had applied customizations in this
unadvised manner, the migration process would succeed but the related
assets bundle was broken.

After this commit
Those customizations are detected and are not migrated.
They are reported in the migration report instead.

Sources on how they are detected
- view key has the form "web_editor.scss_1a2b3c" or something like that
  => see https://github.com/odoo/odoo/blob/74c69537563513bd2399470bbaeceb2018c050c2/addons/web_editor/models/assets.py#L201
- related does not contains "/user_custom_"  and either ends with ".xml"
  either contains ".custom." and ends with ".js" or ".scss"
  => see https://github.com/odoo/odoo/blob/74c69537563513bd2399470bbaeceb2018c050c2/addons/web_editor/static/src/js/common/ace.js#L441
     and https://github.com/odoo/odoo/blob/74c69537563513bd2399470bbaeceb2018c050c2/addons/web_editor/models/assets.py#L119

Linked PR (cancelled) odoo/upgrade#3269

closes odoo/upgrade#3387

Signed-off-by: Christophe Simonis <chs@odoo.com>