[IMP] website/saas~15.3: remove d-none anchors from social media blocks
This commit removes invisible anchors from social media blocks. This
must be done because when the social media block was introduced, it
could leave invisible anchors in the DOM (which is bad), an improvement
of this block has been made so that it does not leave invisible anchors
in the DOM but it does not manage invisible anchors anymore. For this
reason, they must be removed.

See https://github.com/odoo/odoo/pull/85476

task-2739937

closes odoo/upgrade#3292

Related: odoo/odoo#85476
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Co-authored-by: qsm-odoo <qsm@odoo.com>