[IMP] account/saas~15.3.2.1: display cpny's total amount when grouping
Relates to odoo/odoo#87535

When looking at the list of the payments and grouping by company,
the total amount per company is no longer displayed since v15.

This is due to the fact that the new variable used to represent
this data was no longer stored.

By storing this data, we can regain this ability.

task-2801688

closes odoo/upgrade#3401

Signed-off-by: Christophe Simonis <chs@odoo.com>