[IMP] base/saas~14.3: Add custom module assets to report
Ticket#2671541

Add a list of custom modules and their assets that were converted
to IrAssets during the upgrade, to include them in the migration report
with instructions to properly declare them in the manifest.

The report has a link to the odoo 15 doc page about Asset management
that explains all the needed details. It also has a list of the affected
modules and their corresponding assets.

Part-of: odoo/upgrade#3409