[FIX] util/orm: clear env on custom fields as manual context exit
The operations done within the context manager `yield` may leave data in
the towrite/tocompte data structures that can make the call to
`setup_models` fail.

This is especially undesirable on the test crawler that should not fail
on test_prepare/check with uncaught exceptions from previously crawled
views. More precisely, when crawling views, some failing recomputes may
be triggered. If they are not cleared from the environment once the
crawling ends at context manager __exit__ we get an error while setting
up the models. This is because the `tocompute` data structure holds the
field that will trigger the failing compute.

upg-240105
upg-239727

closes odoo/upgrade#3417

Signed-off-by: Alvaro Fuentes <afu@odoo.com>