[FIX] account/saas~14.5: Zero tax line CABA
When creating Zero tax lines, there is an issue for cash basis lines
where two tax line would be created on both the invoice and the cash
basis move.
On the invoice though, only one line should be created, with the cash
basis transition account, while on the cash basis move we should find
the two lines as expected.
Another specificity is that we shouldn't find the tax_line_id and
tax_repartition_line_id on the line on the cash basis move that is
using the cash basis account. It should only be found on the line
using the final account (from the tax repartition line or base line)

Also populate some balances with zero instead of leaving
them empty, as I heard it could cause some issues with specific reports.

closes odoo/upgrade#3450

Signed-off-by: Christophe Simonis <chs@odoo.com>