[FIX] util.convert_field_to_html: convert translations
Fields which are converted to html needs translations to be matched
with source of the field, type would be 'model_terms' now and their
translations would now not use 'p' tag inclusion in it's source
and value. So, removal of it is done to match with the data of the
field.

Closes #3461
Closes #3404

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Christophe Simonis <chs@odoo.com>