[FIX] helpdesk/saas~15.{1,2,3}: update ticket portal visibility ir_rule
When a helpdesk team privacy is set to "portal", portal users
who are followers of a ticket or its team should be able to view that
ticket.

However, currently, only a portal user set as a partner
on the ticket can see it, regardless of the team visibility.

This commit updates the corresponding ir_rule in databases where
it has never been modified to match the desired behavior.

Related:
 - https://github.com/odoo/enterprise/pull/25461
 - https://github.com/odoo/enterprise/pull/26811
 - https://github.com/odoo/enterprise/pull/26832

Closes #3468
Closes #3488

Task-2800363

Signed-off-by: Christophe Simonis <chs@odoo.com>