[IMP] mail/saas~15.4: add a to_delete field on <mail.mail>
Purpose
=======
Improve the performance of mass mailing by removing the <mail.mail>, in
batch, all at once. The unlink operation is very expensive for the ORM,
in terms of SQL queries.

A new field to_delete is required to know which <mail.mail> we need
to delete. The reason is that the failure_type is stored on the
<mail.notification>, and it's possible to have <mail.mail> without
<mail.notification>.

Task-2587345
See odoo/odoo/pull/73271
See odoo/enterprise/pull/24266

closes odoo/upgrade#3481

Related: odoo/odoo#73271
Related: odoo/enterprise#24266
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>