[FIX] mrp/0.0.0: check for zero qty BoMs
From 13.0 up we do not allow zero quantity BoMs. In 13.0 we rely in code
checks to enforce the restriction. Since 14.0[^1] we rely on an SQL
constraint. If we start with a DB in 12.0 we may get to 14.0 without the
constraint in place due to:
```
 odoo.schema: Table 'mrp_bom': unable to add constraint 'mrp_bom_qty_positive' as check (product_qty > 0)
```
Thus allowing for zero quantity in BoMs. This in turn may fail with
ZeroDivisionError[^2] on mrp production operations.

upg-281524

[^1]: https://github.com/odoo/odoo/blob/5df214b1a8160ccdfa73b926f1257d5a41de537c/addons/mrp/models/mrp_bom.py#L80
[^2]: https://github.com/odoo/odoo/blob/f5553c550fc49dac92b4ae69a096c43a9e43a0d1/addons/mrp/models/mrp_production.py#L839

closes odoo/upgrade#3484

Signed-off-by: Christophe Simonis <chs@odoo.com>