[IMP] l10n_in_sale/15.5.1.0: remove l10n_in_shipping_gstin
Always companies only get a tax credit on the bill to GST number(vat),
not on the shipping GST number.

So we remove this field but not the column,
because if a customer still needs data then he can transfer it at his end.

see https://github.com/odoo/odoo/pull/81977
task - 2723111

closes odoo/upgrade#3493

Related: odoo/odoo#81977
Signed-off-by: Christophe Simonis <chs@odoo.com>