[FIX] outlook: fix a traceback when you create a new project
Bug
===
When creating a new project, if the language of Outlook is not en_US,
the project creation failed. The reason for that is that outlook add
the headers "Accept-Language" with the current language, and Odoo try
to load the translation for this language. If it doesn't exist, an
error is raised.

Task-2826471