[REV][FIX] base/saas~14.3: unmigrable web_editor assets
This reverts commit 3fad0a0c70fe651ce6ba3a504fa06af77b634709.

It was not the right approach. It was added because of the following
scenario:
- have a v14.0 website with i.e. a custom navbar_mobile.scss (for which
  the edition displays a warning "Editing a built-in file through this
  editor is not advised, as it will prevent it from being updated during
  future App upgrades.")
- migrate successfully to v15 => an ir.asset record was created to
  replace "navbar_mobile.scss" with something similar to
  "custom.navbar_mobile.scss"
- in the meantime, the original path of the navbar_mobile.scss had
  changed (odoo/odoo#58287) leading to a traceback when trying to load
  the migrated assets bundle...

The idea behind 3fad0a0 was that, as the navbar_mobile.scss file was
edited from the editor AND a warning had been shown doing so, we used
the same conditions (as those for the warning to be shown) and mention
it in the migration reports without migrating it.

But this brought an unwanted side effect: any customizations from the
web_editor that was not "/user_custom_..." i.e. change one of the 5 main
colors (user_color_palette) was reported as un-migratable.

A better approach is to provide migration scripts to retarget existing
ir.asset records everytime a file is moved/renamed.

Part-of: odoo/upgrade#3506