[FIX] base/0.0.0: always cleanup records before adding foreign key
As the call can be in a subtransaction (savepoint), we cannot commit the
current transaction. To avoid adding a new subtransaction, we invert the
logic and cleanup the table *before* we even try to add the constraint
instead of doing it only if it failed.

Fix queries to ignore rows where the FK column is NULL.

closes odoo/upgrade#3519

Signed-off-by: Christophe Simonis <chs@odoo.com>