[IMP] website/saas~14.5.1.0: Adapt Contactus views to company data
In Odoo v14, `the contactus` and `contactus-thank-you` pages display
the company description which are "dynamic" and the data are taken from
the database.
Since Odoo saas~14.5, these page are static. It means that the data has
to be maually modified to fit the company description (to avoid
displaying the default data).

This is why we could set up the description on the two pages, on each
websites (matching the right company).

closes odoo/upgrade#3538

Signed-off-by: Christophe Simonis <chs@odoo.com>