[IMP] project/saas~15.4: enable milestone feature if a milestone exists
Before this commit, a new group is added to globally enable the
"Milestone" feature in the Project app if the internal users has this
group called `group_project_milestone`. By default, the feature is
disabled but since the milestone model exists from v.15, the feature
has to be enabled in the dbs in which a milestone is created in at
least one project.

This commit globally enables the feature if a milestone is created in
at least a project in the db. Also, it will enable the feature on the
existing projects in which a milestone is created.

Community PR: odoo/odoo#90972
Enterprise PR: odoo/enterprise#27146

task-2829542

Part-of: odoo/upgrade#3550