[IMP] sale_project/saas~15.4: convert milestone product into manual one
Before this commit, when the "Milestones" feature is globally
disabled, it will no longer be possible to create a milestone product.
Moreover, if the user wants to see the form view of a milestone
products, it will crash because the 'delivered_milestones' is removed in
`service_policy` field if the feature is disabled. For the Sale Order
Items containing a milestone product, the `qty_delivered_method` has to
be updated to 'manual' to allow the user manually editing the quantity
delivered (`qty_delivered` field) otherwise it will be impossible for
the user to update this field.

This commit checks if the feature is globally disabled, if it is the
case then the milestone products will be convert into delivered manual
service product and the quantity delivered method for the Sales Order
Items containing those products will be updated to 'manual'.

Community PR: odoo/odoo#90972
Enterprise PR: odoo/enterprise#27146

task-2829542

closes odoo/upgrade#3550

Signed-off-by: Christophe Simonis <chs@odoo.com>