[FIX] account/saas~13.4: fix inconsistent flush
`cr.commit()` will flush the environment before performing the commit.
Since it does not have an environment, it will look for one that uses
the same cursor and flush it. If there is more than one, it will not
choose the same every time. Because each environment comes with its own
context, the flush will not always be done with the same context and
will yield inconsistent results.

This commit forces a flush with the correct context before committing.

closes odoo/upgrade#3549

Signed-off-by: Christophe Simonis <chs@odoo.com>