[FIX] sale_timesheet_enterprise/saas~15.4: fix version
As the sale_timesheet_enterprise is in version 0.1, it seems that there
was a mistake in the directory naming.

Also, as an action is removed with `remove_view`, the script crashes.
So the whole function is commented and has to be fixed later in order to
allow bumping Odoo master version.

closes odoo/upgrade#3571

Related: odoo/odoo#92997
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>