[ADD] base/saas~15.2.1.3:rm duplicate partner_bank
...s before applying unique constraint

See https://github.com/odoo/odoo/pull/81362

A constraint on the ResPartnerBank model changes between saas-15.1 and
saas-15.2. When applying the upgrade, the constraint might not be
applied in the case of some multicompany databases.

The reason is that the old constraint is:
unique(sanitized_acc_number, company_id)
allowing for multiple 'same' accounts on the same partner, provided the
record was created by different companies.
Then constraint is:
unique(sanitized_acc_number, partner_id)
which can conflict with the aforementioned configuration.

The solution in this commit locates all duplicate res_partner_bank
records by the standard of the new constraint, and from among them
selects the most pertinent record, with the most relevant data, to be
the single res_partner_bank. The other records are de-referenced and
deleted.

The company ids of the new, singular res_partner_bank is set to the
company_id of the partner.

closes odoo/upgrade#3583

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Christophe Simonis <KangOl@users.noreply.github.com>