[IMP] gmail: use restricted scope to fetch data
Purpose
=======
Use a more restrictive scope to use the minimum we need for the
application.

With this restricted scope, we can only get once the Email object. This
object is not serializable, so we can not give it to the next card
when the user do a search.

To counter this issue, we use a second access token (which can be
serialized) to get the email attachments.

Ask the email body only if necessary.

Task-2745848