[IMP] survey/saas~15.5: set correct question_type for page questions
Currently, in demo data (or data) of some surveys, there are records of
'survey.question' having 'is_page' field value as true and 'question_type'
is not defined. When question type is not defined, default value
'simple_choice' is assigned to the 'question_type' field. Now when we
create a live session of that survey and goes to the page, it tries to read
the data for bar chart but data is not available (undefined) and throws the
traceback. This happens  because question types 'simple_choice' and
'multiple_choice') are expected to load a bar chart.

Linked PRs updates the data/demo data and adds constraint/compute
methods to to make sure that for pages, the `question_type` is set to
False. And to adapt the historic data, this script updates the pages
and sets the 'question_type' to NULL so avoid the traceback.

Community PR: https://github.com/odoo/odoo/pull/93414
Enterprise PR: https://github.com/odoo/enterprise/pull/28305

taskID-2841582

closes odoo/upgrade#3600

Related: odoo/odoo#93414
Related: odoo/enterprise#28305
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>