[IMP] website/14.0.1.0: Adapt footer data
In Odoo 13.0, the footer data are taken from the database.
In Odoo 14.0, the data are static and the data are back to the default
ones (set direclty in the template).

To avoid customer to have to manually set the data on the footer,
a script would take every footer view and adapt it with the right company
details (for each websites)

closes odoo/upgrade#3616

Signed-off-by: Christophe Simonis <chs@odoo.com>