[FIX] base/14.0: On delete restrict for m2o ir models
Since v14[^1], fields relating to ir models cannot have their on delete set
to `restrict`.

It should raise an error when that happens, but for older custom modules that
might use that, it's necessary to change the value to avoid the field from not
loading and breaking the upgrade if it's used somewhere, e.g. test mock crawl.

This fixes the `audit.rule.model_id` that has been fixed several times
in the specific repository.

upg-230918
upg-360409

[^1]: https://github.com/odoo/odoo/pull/38847

closes odoo/upgrade#3630

Signed-off-by: Christophe Simonis <chs@odoo.com>