[IMP] account_asset/saas~16.1: remove a field and rename others
The field `linked_asset_type` is not used anymore so it is removed
We rename the field `draft_asset_ids` to `draft_asset_exist`
as it is a boolean and not some ids.
We rename the field `number_asset_ids` to `count_asset` as it is
a count and not some ids.

taskid-2894690

see odoo/enterprise#28926

closes odoo/upgrade#3634

Signed-off-by: Christophe Simonis <chs@odoo.com>